// src/app/shared/models/stakeholder/request/stakeholderEditRequest.ts
import { ContactAddRequest } from './../../contact/request/contactaddrequest';
import { ContactEditRequest } from '../../contact/request/contactEditRequest';

export class StakeholderEditRequest{
    address: string;
    // amphurCode: string;
    postCode:string;
    // provinceCode:string;
    stakeholderCode: string;
    stakeholderName: string;
    // tambolCode: string;
    taxId: string;
}